requirements.txt
# Core Framework

# Use Python 3.12 or 3.13 for best compatibility
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
pydantic-settings==2.0.3

# Redis/Cache (Optional)
redis==5.0.1
upstash-redis==0.15.0

# Azure Blob Storage
azure-storage-blob==12.19.0

# LLM & AI
openai==1.54.0
anthropic==0.18.1
tiktoken==0.5.2

# Document Processing
pypdf==4.0.1
pdfplumber==0.10.3
beautifulsoup4==4.12.3
lxml==5.1.0
python-magic==0.4.27

# HTTP & Async
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Azure Storage
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
jinja2==3.1.3
python-dateutil==2.8.2
loguru==0.7.2
pytz==2024.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
